+++ NEWS.md
@@ -1,3 +1,3 @@
-# RootScanR 1.0.0
+# Rootopia 1.0.0
 
 * Initial CRAN submission.